backend/requirements.txt
flask
tensorflow
numpy
pillow
flask-cors
